fix:rollback - actions
